ajustess

+++ MonkeyDelivery/Src/Control/States/CreditsState.cpp
@@ -9,8 +9,8 @@
 CreditsState::CreditsState(Game* game) : State(game)
 {
 
-	buttonW = game->getWindowWidth() / 6;
-	buttonH = game->getWindowHeight() / 7.143;
+	buttonW = (int)game->getWindowWidth() / 6;
+	buttonH = (int)(game->getWindowHeight() / 7.143);
 
 	addButton(new Back1((int)game->getWindowWidth() / 2 - buttonW/2, (int)game->getWindowHeight() - 250, buttonW, buttonH, game));
 	backgroundTexture = game->getTexture(bckg_options);
